Compressed p3d format
Normal compression 64,32,16,5 bit
p3d.exe support for compressed models

+++ include/pr/view3d/util/wrappers.h
@@ -101,12 +101,12 @@ namespace pr
 				// Just get the description.
 				buf->GetDesc(this);
 			}
-			BufferDesc(size_t count, size_t element_size_in_bytes, EUsage usage = EUsage::Default, EBind bind_flags = EBind::UnorderedAccess, ECPUAccess cpu_access = ECPUAccess::None, EResMisc res_flag = EResMisc::None)
+			BufferDesc(size_t count, int element_size_in_bytes, EUsage usage = EUsage::Default, EBind bind_flags = EBind::UnorderedAccess, ECPUAccess cpu_access = ECPUAccess::None, EResMisc res_flag = EResMisc::None)
 				:BufferDesc()
 			{
 				Init(count, element_size_in_bytes, nullptr, usage, bind_flags, cpu_access, res_flag);
 			}
-			BufferDesc(size_t count, void const* data, size_t element_size_in_bytes, EUsage usage = EUsage::Default, EBind bind_flags = EBind::UnorderedAccess, ECPUAccess cpu_access = ECPUAccess::None, EResMisc res_flag = EResMisc::None)
+			BufferDesc(size_t count, void const* data, int element_size_in_bytes, EUsage usage = EUsage::Default, EBind bind_flags = EBind::UnorderedAccess, ECPUAccess cpu_access = ECPUAccess::None, EResMisc res_flag = EResMisc::None)
 				:BufferDesc()
 			{
 				Init(count, element_size_in_bytes, data, usage, bind_flags, cpu_access, res_flag);
@@ -121,7 +121,7 @@ namespace pr
 			{
 				Init(Sz, sizeof(Elem), data, usage, bind_flags, cpu_access, res_flag);
 			}
-			void Init(size_t count, size_t element_size_in_bytes, void const* data, EUsage usage, EBind bind_flags, ECPUAccess cpu_access, EResMisc res_flag)
+			void Init(size_t count, int element_size_in_bytes, void const* data, EUsage usage, EBind bind_flags, ECPUAccess cpu_access, EResMisc res_flag)
 			{
 				Data                = data;                                // The initialisation data (or null)
 				ElemCount           = count;                               // The number of elements in the buffer
@@ -145,7 +145,7 @@ namespace pr
 			VBufferDesc()
 				:BufferDesc()
 			{}
-			VBufferDesc(size_t count, size_t element_size_in_bytes, EUsage usage = EUsage::Default, ECPUAccess cpu_access = ECPUAccess::None, EResMisc res_flag = EResMisc::None)
+			VBufferDesc(size_t count, int element_size_in_bytes, EUsage usage = EUsage::Default, ECPUAccess cpu_access = ECPUAccess::None, EResMisc res_flag = EResMisc::None)
 				:BufferDesc(count, element_size_in_bytes, usage, EBind::VertexBuffer, cpu_access, res_flag)
 			{}
 			template <typename Elem> VBufferDesc(size_t count, Elem const* data, EUsage usage = EUsage::Default, ECPUAccess cpu_access = ECPUAccess::None, EResMisc res_flag = EResMisc::None)
@@ -169,11 +169,11 @@ namespace pr
 				:BufferDesc()
 				,Format(DXGI_FORMAT_UNKNOWN)
 			{}
-			IBufferDesc(size_t count, size_t element_size_in_bytes, DXGI_FORMAT format, EUsage usage = EUsage::Default, ECPUAccess cpu_access = ECPUAccess::None, EResMisc res_flag = EResMisc::None)
+			IBufferDesc(size_t count, int element_size_in_bytes, DXGI_FORMAT format, EUsage usage = EUsage::Default, ECPUAccess cpu_access = ECPUAccess::None, EResMisc res_flag = EResMisc::None)
 				:BufferDesc(count, element_size_in_bytes, usage, EBind::IndexBuffer, cpu_access, res_flag)
 				,Format(format)
 			{}
-			IBufferDesc(size_t count, void const* data, size_t element_size_in_bytes, DXGI_FORMAT format, EUsage usage = EUsage::Default, ECPUAccess cpu_access = ECPUAccess::None, EResMisc res_flag = EResMisc::None)
+			IBufferDesc(size_t count, void const* data, int element_size_in_bytes, DXGI_FORMAT format, EUsage usage = EUsage::Default, ECPUAccess cpu_access = ECPUAccess::None, EResMisc res_flag = EResMisc::None)
 				:BufferDesc(count, data, element_size_in_bytes, usage, EBind::IndexBuffer, cpu_access, res_flag)
 				,Format(format)
 			{}
